+++ examples/typescript/CrowdFunding.ts
@@ -14,19 +14,25 @@ interface CrowdFundingSchema extends BaseContractSchema {
 	settle: () => Promise<Transaction>;
 }
 
+const contractPath = '../assets/CrowdFunding.sol';
+const contractFile = fs.readFileSync(contractPath, 'utf8');
+const contractName = ':' + 'CrowdFunding';
+const output = solc.compile(contractFile, 1);
+const compiled = {
+	bytecode: output.contracts[contractName].bytecode,
+	abi: JSON.parse(output.contracts[contractName].interface)
+};
+
 const from = '0X5E54B1907162D64F9C4C7A46E3547084023DA2A0'.toLowerCase();
+const directory = new DataDirectory('/Users/danu/.evmlc');
+const account = directory.keystore.decryptAccount(from, 'asd');
+const contractAddress = '0x38CB86c8123e68164390259D022b5D2afffCB273';
+
 const evmlc = new EVMLC('127.0.0.1', 8080, {
 	from,
 	gas: 1000000,
 	gasPrice: 0
 });
-const directory = new DataDirectory('/Users/danu/.evmlc');
-const account = directory.keystore.decrypt(from, 'asd');
-const contractAddress = '0x38CB86c8123e68164390259D022b5D2afffCB273';
-const compiled = evmlc.compileContract(
-	'CrowdFunding',
-	'../assets/CrowdFunding.sol'
-);
 
 const loadContract = async () => {
 	return await evmlc.loadContract<CrowdFundingSchema>(compiled.abi, {