+++ test/setup.ts
@@ -1,12 +1,11 @@
-// @ts-ignore
-import shell from 'shelljs'
-
-import {exec, spawn} from 'child_process';
+import * as Chai from 'chai';
 
 import { EVMLC } from '../src';
 
+export const assert = Chai.assert;
+
 const evmlc = new EVMLC('127.0.0.1', 8080, {
-	from: '0X5E54B1907162D64F9C4C7A46E3547084023DA2A0'.toLowerCase(),
+	from: '0X5E54B1907162D64F9C4C7A46E3547084023DA2A0',
 	gas: 100000,
 	gasPrice: 0
 });